requirements.txt
# General Python Libraries
pandas==2.0.1
schedule==1.1.0
boto3==1.26.99
azure-mgmt-compute==29.0.0
azure-identity==1.12.0
google-cloud-compute==1.7.0
google-auth==2.19.1

# Web Framework
Flask==2.3.2

# Testing Libraries
unittest-xml-reporting==3.2.0
pytest==7.3.0
requests==2.28.1

python-dotenv==1.0.0